refactor: move types/result to types/util
