+++ v-next/hardhat/package.json
@@ -38,7 +38,6 @@
     "./types/solidity": "./dist/src/types/solidity.js",
     "./console.sol": "./console.sol",
     "./utils/contract-names": "./dist/src/utils/contract-names.js",
-    "./types/result": "./dist/src/types/result.js",
     "./utils/result": "./dist/src/utils/result.js",
     "./types/runtime": "./dist/src/internal/deprecated-module-imported-from-hardhat2-plugin.js",
     "./builtin-tasks/task-names": "./dist/src/internal/deprecated-module-imported-from-hardhat2-plugin.js",
