+++ v-next/hardhat/src/types/utils.ts
@@ -50,3 +50,10 @@ export type Return<T> = T extends (...args: any[]) => infer Ret ? Ret : never;
 export type RequireField<T, K extends keyof T> = T & {
   [P in K]-?: T[P];
 };
+
+/**
+ * A result that can either have a value or an error.
+ */
+export type Result<ValueT, ErrorT> =
+  | { readonly success: true; readonly value: ValueT }
+  | { readonly success: false; readonly error: ErrorT };
